requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
python-multipart
pillow
libzbar0
pdf2image
poppler-utils
